+++ cmd/sec.go
@@ -34,7 +34,6 @@ func init() {
 		newCmdObjectMonitor(kind),
 		newCmdObjectStatus(kind),
 		newCmdObjectUnset(kind),
-		newCmdSecFullPEM(kind),
 		newCmdSecGenCert(kind),
 		newCmdSecPKCS(kind),
 	)